Fix name of LICENSE file in MANIFEST.in
This avoids a warning during build (and during pip install).

Also, fix license parameter in setup.py. This param should
be the name of the license, not the file containing the
license. Since this package is dual licensed I copied the
syntax used in setuptools itself which is "XXX or YYY".