Fixes multiprocess inconsistency and adds inline-only option.
When running with verbosity >= 2, test times get appended
to nose's own test output - (e.g.:
my_test.py:MyTestCase.test_a ... ok (0.1653s)) -
but this isn't happening when the multiprocess plugin is used.
This commit fixes that.

Additionally adds an option to rely on that output exclusively,
and suppress the extra line per test that outputs the same
information.